webapi/webapi-messaging-sysapps-tests/messaging/MessagingMnanger_method_message.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
may be used to endorse or promote products derived from this work without
specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: MessagingManager-method-message</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#messagingmanager-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var filter = {
  type: "sms"
};
var options = {
  sortBy: "date",
  sortOrder: "descending",
  limit : 3
};
var messageID = "0123456";

var t1 = async_test("Check if the findMessages methos of message can return a MessagingCursor object", {timeout: 4000});
var t2 = async_test("Check if the getMessage methos of message can return a message object", {timeout: 4000});
var t3 = async_test("Check if the markMessageRead of message can change message status", {timeout: 4000});
var t4 = async_test("Check if the markMessageRead of message which has three params can change message status", {timeout: 4000});
var t5 = async_test("Check if the deleteMessage of message can delete the message by messageID", {timeout: 4000});

t1.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, options).then(function (msg) {
    t1.step(function () {
      assert_true(!!msg, "the return value should be MessagingCursor");
      if (msg.element.lenght > 0) {
        messageID = msg.element[0].messageID;
      }
    });
    t1.done();
  }, function (error) {
    t1.step(function () {
      assert_unreached(error.message);
    });
    t1.done();
  });
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.getMessage(messageID).then(function (data) {
    t2.step(function () {
      assert_true(data instanceof SmsMessage || data instanceof MmsMessage, "the return value should be smsmessage or mmsmessage");
    });
    t2.done();
  }, function (error) {
    t2.step(function () {
      assert_unreached(error.message);
    });
    t2.done();
  });
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.markMessageRead(messageID, true).then(function (data){
    t3.step(function () {
      assert_equals(data.messageID, "true", "the return value should be true");
      message.getMessage(messageID).then(function (msg) {
        assert_true(msg.read, "this message's read status should be true");
      }, function (error) {
        assert_unreached("the getMessage function run exception");
      });
    });
    t3.done();
  }, function (error) {
    t3.step(function () {
      assert_unreached(error.message);
    });
    t3.done();
  });
});

t4.step(function () {
  assert_true(!!message, "The define of message");
  message.markMessageRead(messageID, false, true).then(function (data){
    t4.step(function () {
      assert_equals(data.messageID, "false", "the return value should be false");
      message.getMessage(messageID).then(function (msg) {
        assert_true(!msg.read, "this message's read status should be false");
      }, function (error) {
        assert_unreached("the getMessage function run exception");
      });
    });
    t4.done();
  }, function (error) {
    t4.step(function () {
      assert_unreached(error.message);
    });
    t4.done();
  });
});

t5.step(function () {
  assert_true(!!message, "The define of message");
  message.deleteMessage(messageID).then(function (data){
    t5.step(function () {
      assert_true(data instanceof boolean, "the return value should be defined");
      message.getMessage(messageID).then(function (msg) {
        assert_false(data instanceof SmsMessage || data instanceof MmsMessage, "getMessage method should not find this message");
      });
    });
  }, function (error) {
    t5.step(function () {
      assert_unreached(error.message);
    });
    t5.done();
  });
});

</script>
